scheduler-service/src/requirements.txt
fastapi
uvicorn[standard]
httpx
croniter
lmdb
psycopg2
aiokafka
aioredis
sqlalchemy
dataclasses-json
SQLAlchemy-Utils
email-validator
strawberry-graphql[fastapi]
pyjwt
pytz